feat: spend reducers and tests
