+++ src/vms/pvm/etna-builder/spend-reducers/errors.ts
@@ -1,0 +1,3 @@
+export const IncorrectStakeableLockOutError = new Error(
+  'StakeableLockOut transferOut must be a TransferOutput.',
+);
