+++ src/vms/pvm/etna-builder/spend-reducers/index.ts
@@ -1,0 +1,6 @@
+export { handleFeeAndChange } from './handleFeeAndChange';
+export { useSpendableLockedUTXOs } from './useSpendableLockedUTXOs';
+export { useUnlockedUTXOs } from './useUnlockedUTXOs';
+export { verifyAssetsConsumed } from './verifyAssetsConsumed';
+
+export type * from './types';
