+++ src/vms/pvm/etna-builder/spend-reducers/types.ts
@@ -1,0 +1,13 @@
+import type { Context } from '../../../context';
+import type { SpendProps } from '../spend';
+import type { SpendHelper } from '../spendHelper';
+
+export type SpendReducerState = Readonly<
+  Required<Omit<SpendProps, 'shouldConsolidateOutputs'>>
+>;
+
+export type SpendReducerFunction = (
+  state: SpendReducerState,
+  spendHelper: SpendHelper,
+  context: Context,
+) => SpendReducerState;
